+++ Runtime/Base/NotchSolutionUtility.cs
@@ -18,16 +18,30 @@ namespace E7.NotchSolution
         private const string widestAspectIndex = prefix + nameof(widestAspectIndex);
 
         /// <summary>
-        /// A smart accessor which returns simulated relative safe area in-editor, or a real one outside of editor.
+        /// A smart accessor which returns Notch Solution simulated relative safe area in-editor,
+        /// hijacked safe area and other device information in 2019.3 with Unity Device Simulator package,
+        /// or a real one outside of editor.
         /// </summary>
         public static Rect SafeAreaRelative
         {
             get
             {
 #if UNITY_EDITOR
-                // Need the simulator to calculate this for us.
-                return NotchSolutionUtility.SimulatedSafeAreaRelative;
-#else
+                bool useNotchSimulator = true;
+#if UNITY_2019_3_OR_NEWER
+                if (UnityDeviceSimulatorActive)
+                {
+                    //Trust the Unity-modified `Screen.safeArea`, do the same thing as runtime.
+                    useNotchSimulator = false;
+                }
+#endif
+                if (useNotchSimulator)
+                {
+                    // Need the simulator to calculate this for us.
+                    return NotchSolutionUtility.SimulatedSafeAreaRelative;
+                }
+#endif
+
                 Rect absolutePaddings = Screen.safeArea;
                 return new Rect(
                     absolutePaddings.x / Screen.width,
@@ -35,12 +49,31 @@ namespace E7.NotchSolution
                     absolutePaddings.width / Screen.width,
                     absolutePaddings.height / Screen.height
                 );
-#endif
             }
         }
 
-#if UNITY_EDITOR
+#if UNITY_2019_3_OR_NEWER
+        static System.Type T = System.Type.GetType("UnityEditor.PlayModeView,UnityEditor");
+        static System.Reflection.MethodInfo GetMainPlayModeView = T.GetMethod("GetMainPlayModeView", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Static);
 
+        /// <summary>
+        /// With [Device Simulator](https://docs.unity3d.com/Packages/com.unity.device-simulator@latest) installed it can use a secondary
+        /// game view (also a new internal feature in 2019.3) and take control of the screen size and etc. If we detect that active, disable our
+        /// own overrides and just go along with Device Simulator simulated value.
+        /// </summary>
+        private static bool UnityDeviceSimulatorActive
+        {
+            get
+            {
+                var mainPlayModeView = GetMainPlayModeView.Invoke(null,null);
+                var name = mainPlayModeView.GetType().FullName;
+                //I am lazy so I will simply do a class name check with the one in that package.
+                return name == "Unity.DeviceSimulator.SimulatorWindow";
+            }
+        }
+#endif
+
+#if UNITY_EDITOR
         public static (bool landscapeCompatible, bool portraitCompatible) GetOrientationCompatibility()
         {
             bool landscapeCompatible = false;